First Section is Created
I created first section with header(company logo and name; navigation links; login and sign up buttons), company moto, text about them and get-started button. All of this are with a background video, with play button on the right. Box is created on the bottom with history stats and text. Section is responsive and is adjustable to all dimensions. Video changes to photo for smaller dimensions and watch-video button appears. Also font-family is added to all text tags: p, h1, h2, h3 and cursor pointers are added to all links and buttons. Media and icon files are added, too.